Deprecate `tag` with positional arguments
Modern evolutions of the `ActionView::TagHelper#tag` method are intended
to be called _on_ the `TagBuilder` instance itself instead of with
positional arguments.

For example, `tag("div")` would be called with `tag.div`.

This commit adds a deprecation warning to encourage callers to invoke
the helper in the more modern style.

Internally, Rails utilizes the positional argument syntax in several
places. Since deprecations must be part of minor version releases, this
commit wraps the internal invocations in `ActionView.deprecator.silence`
blocks so that they can remain while the deprecation is made public.
This compromise buys the framework time to re-structure those
invocations to be compatible with the more modern syntax before the next
major release.

There are several calls to `tag` with positional arguments in the test
suite. This commit wraps those as well. In the future, some of those
tests will be able to remove those blocks, while other tests will be
able to be removed entirely once that style of invocation is no longer
supported.