`form_tag` and `form_with`: close tag when block omitted
Prior to this change, calling a `form_tag` or `form_with` without
passing a block generates a `<form>` element without a closing `</form>`
tag.

The [form_for][] version of building a `<form>` element requires a block
to execute, so omitting the closing `</form>` element isn't a
possibility.

A current work-around to achieve the desired behavior is to declare the
`form_tag` or `form_with` call with an empty block:

```erb
<%= form_tag "https://example.com" do %>
<% end %>

<%= form_with url: "https://example.com" do %>
<% end %>
```

Unfortunately, being unaware of this issue will likely yield HTML that
is silently invalid, which can cause surprising nesting and behavior,
such as elements declared after the `<form>` becoming descendants
instead of siblings.

Introduce the `config.action_view.closes_form_tag_without_block`
configuration value.

By default (in versions prior to 7.2), that value is `false` and
preserves backwards compatibility. When set to `true`, calls to
`form_with` and `form_tag` without a block will result in elements that
close themselves.

[form_for]: https://github.com/rails/rails/blob/efae65d005ee298207d720fb4f61c28a38973e8e/actionview/test/template/form_helper_test.rb#L1555-L1560