Add cost argument to has_secure_password
There may be instances where a developer is working in an environment where a higher cost than the default (12) for BCrypt is required either due to arbitrary security requirements or the recommended minimum changes and the version of bcrypt used isn't updated.

This change adds a `cost:` argument to `has_secure_password`. If provided, it will use that value for the password cost. Its value is validated to make sure it's between `BCrypt::Engine::MIN_COST` and `BCrypt::Engine::MAX_COST`.

Additionally `[password attribute]_cost` is exposed as a convenience method to determine the cost used for the set password hash. This can be used to review the current costs from a user record and either transparently upgrade the user's password or force a password reset.