Fix label with `for` option not getting prefixed by form `namespace` value
When a `form_for` has the `namespace` option, all other fields are
prefixed with it except for the `label` tag which has the `for`
option.

Example:
```rb
form_for(@post, namespace: SecureRandom.urlsafe_base64) do |f|
  concat f.label(:title, for: "my_custom_post_title")
  concat f.text_field(:title, id: "my_custom_post_title")
end
```
In the above example, the label will not be linked to the text field,
which shouldn't be the case just because the `namespace` option is
present.